Move mapState to core/foundation/common and fix replayCache
- Move StateFlowExtensions.kt from lib/design/fe to core/foundation/common
- Delegate replayCache to upstream's replayCache with transform
