+++ core/foundation/common/src/commonMain/kotlin/cz/adamec/timotej/snag/core/foundation/common/StateFlowExtensions.kt
@@ -10,7 +10,7 @@
  * Department of Software Engineering
  */
 
-package cz.adamec.timotej.snag.lib.design.fe.state
+package cz.adamec.timotej.snag.core.foundation.common
 
 import kotlinx.coroutines.flow.FlowCollector
 import kotlinx.coroutines.flow.StateFlow
@@ -23,7 +23,7 @@ fun <T, R> StateFlow<T>.mapState(transform: (T) -> R): StateFlow<R> =
     object : StateFlow<R> {
         override val value: R get() = transform(this@mapState.value)
 
-        override val replayCache: List<R> get() = listOf(value)
+        override val replayCache: List<R> get() = this@mapState.replayCache.map(transform)
 
         override suspend fun collect(collector: FlowCollector<R>): Nothing {
             this@mapState.collect { collector.emit(transform(it)) }
